Parse shader inputs from SpirV (#42)
* Detect shader kind from spir-v contents

Instead of based on naming convention.

* Improve asset-loading error handling

* Read shader entrypoint name from spir-v

* Refactor graphic to have list of shaders

Reason is to be more flexible in the future when
we want to support more stader stages.

* Read shader inputs from SpirV

* Refactor desc manager to allow flexible order

Instead of fixed binding order of resources you
can now just specify the resource type for a binding
id. This also allows for graps in the binding numbers.

* Bind textures based on shader resources

Allows using non-sequential bindings (which can
help in shader re-use).